requirements.txt
# HRP Core Dependencies
# Install: pip install -r requirements.txt

# Data
duckdb>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Backtesting
vectorbt>=0.26.0
empyrical-reloaded>=0.5.10  # Portfolio performance metrics (battle-tested)

# ML
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
optuna>=3.4.0
hmmlearn>=0.3.0  # For HMM regime detection

# Experiment tracking
mlflow>=2.10.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Agent integration
fastmcp>=0.1.0
anthropic>=0.18.0

# Scheduling & notifications
apscheduler>=3.10.0
resend>=0.7.0

# Data sources
yfinance>=0.2.30
polygon-api-client>=1.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
httpx>=0.26.0
exchange-calendars>=4.5.0
